Calculation of ranking

+++ app/Http/Controllers/WorkController.php
@@ -24,14 +24,14 @@ class WorkController extends Controller {
 			'more' => [
 				'idMore' => 0,
 				'name' => false,
-				'value' => 0,
+				'value' => '0:0',
 				'email' => '',
 				'department' => '',
 			],
 			'less' => [
 				'idLess' => 0,
 				'name' => false,
-				'value' => 0,
+				'value' => '0:0',
 				'email' => '',
 				'department' => '',
 			],
@@ -39,12 +39,14 @@ class WorkController extends Controller {
 
 		foreach ($works as $work) {
 
+			$work->hours = formatHour($work->hours, $work->minutes);
+
 			foreach ($departments as $department) {
 
 				if ($work->idDepartment == $department->id) {
 					$work->idDepartment = $department->name;
 
-					if ($work->hours > $positions['more']['value']) {
+					if (str_replace(':', '.', $work->hours) > str_replace(':', '.', $positions['more']['value'])) {
 						$positions['more']['value'] = $work->hours;
 						$positions['more']['idLess'] = $work->id;
 						$positions['more']['name'] = $work->name;
@@ -52,7 +54,7 @@ class WorkController extends Controller {
 						$positions['more']['department'] = $work->idDepartment;
 					}
 
-					if ($work->hours < $positions['less']['value']) {
+					if (str_replace(':', '.', $work->hours) < str_replace(':', '.', $positions['less']['value'])) {
 						$positions['less']['value'] = $work->hours;
 						$positions['less']['idLess'] = $work->id;
 						$positions['less']['name'] = $work->name;
@@ -62,7 +64,6 @@ class WorkController extends Controller {
 
 				}
 
-				$work->hours = formatHour($work->hours, $work->minutes);
 			}
 
 		}
@@ -85,9 +86,15 @@ class WorkController extends Controller {
 
 	public function remove($id) {
 		$work = Work::find($id);
-		$work->delete();
+		$records = Record::where('idWork', $id)->count();
 
-		return flashMessage('Work', 'Funcionário removido com sucesso');
+		if ($records == 0) {
+			$work->delete();
+			return flashMessage('Work', 'Funcionário removido com sucesso');
+		} else {
+			return flashMessage('Work', 'Este funcionário possui registros, será necessário apagar antes de excluí-lo', 'info');
+		}
+
 	}
 
 	public function novo() {
